Add Makefile and enableEmoji to hugo.toml, update my-first-post.md, and spark-excel.md
---

Added Makefile for running and building the project.
Updated the 'my-first-post.md' with an introduction to the first post on the site and the use of Hugo framework for building the site.
Updated 'spark-excel.md' file with a code example for reading and loading excel files using the Spark Excel library.
Also added enableEmoji to the hugo.toml configuration file.

---